`Rename router to routes and remove deleted file`
